tick-double

+++ src/hotspot/share/jfr/periodic/jfrThreadCPULoadEvent.cpp
@@ -132,7 +132,7 @@ void JfrThreadCPULoadEvent::send_events() {
     }
   }
   log_trace(jfr)("Measured CPU usage for %d threads in %.3f milliseconds", number_of_threads,
-    (double)(JfrTicks::now() - event_time).milliseconds());
+    (JfrTicks::now() - event_time).milliseconds());
   // Restore this thread's thread id
   periodic_thread_tl->set_thread_id(periodic_thread_id);
 }
